backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
python-multipart==0.0.12
python-dotenv==1.0.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
